+++ tests/testthat/helper-write_parameter_file.R
@@ -5,15 +5,18 @@ write_sample_parameters_file <- function(value,
                                          disease,
                                          parameter,
                                          start_date,
-                                         end_date) {
+                                         end_date,
+                                         geo_value,
+                                         reference_date) {
   Sys.sleep(0.05)
   df <- data.frame(
     start_date = as.Date(start_date),
-    geo_value = state,
     disease = disease,
     parameter = parameter,
     end_date = end_date,
-    value = I(list(value))
+    geo_value = geo_value,
+    value = I(list(value)),
+    reference_date = reference_date
   )
 
   con <- DBI::dbConnect(duckdb::duckdb())